+++ respa_admin/views/_login.py
@@ -5,7 +5,7 @@ from django.contrib import auth as django_auth
 from django.http import HttpResponseBadRequest, HttpResponseRedirect
 from django.urls import reverse
 from django.views.generic.base import TemplateView
-from tkusers.providers.turku_oidc.views import TurkuOIDCOAuth2Adapter
+from tkusers.providers.tunnistamo.views import TurkuOIDCOAuth2Adapter
 from respa_admin.views.base import ExtraContextMixin
 
 from ..auth import is_allowed_user
@@ -61,7 +61,7 @@ def tunnistamo_login(request):
         if not is_allowed_user(request.user):
             return _logout_locally_and_in_tunnistamo(request)
 
-    url = _get_url_with_next(request, 'turku_oidc_login')
+    url = _get_url_with_next(request, 'tunnistamo_login')
     return HttpResponseRedirect(url)
